% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootreg.R
\name{bootreg}
\alias{bootreg}
\title{Generate Johnson-Neyman plot}
\usage{
bootreg(data, model, method, n = 1000)
}
\arguments{
\item{data}{a data frame}

\item{model}{a fitted lme4-model}

\item{method}{a function to use for refitting}

\item{n}{number of bootstrap samples}
}
\value{
Bootstrap regression results object (brr)
}
\description{
Generates a Johnson-Neyman plot for the effect of X on the dependent variable, moderated by M. Performs two-sided t-Test on coefficient.
}
\examples{
#mixed models
x <- rnorm(50)
m <- rnorm(50)
y <- rnorm(50,x*m,2)
g <- sample(1:10,50,TRUE)
model <- lme4::lmer(y~m*x + (1|g))
brr <- bootreg(data.frame(x,m,y,g),model,lme4::lmer)

}
