% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnson_neyman.R
\name{.johnson_neyman}
\alias{.johnson_neyman}
\title{Generate Johnson-Neyman plot}
\usage{
.johnson_neyman(B, sigma, x, m, modrange, df, resolution = 10000)
}
\arguments{
\item{B}{coefficient vector}

\item{sigma}{covariance matrix}

\item{x}{name of IV}

\item{m}{name of M}

\item{df}{df for t-test}

\item{resolution}{step count for solver}
}
\value{
Johnson-neyman plot
}
\description{
Internal function to generate johnson-neyman plot given parameters
}
