% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelType.R
\name{.modelType}
\alias{.modelType}
\title{Find out what type the model is}
\usage{
.modelType(object)
}
\arguments{
\item{object}{model object}
}
\value{
string specifying package the model originates frmo
}
\description{
Find out what type the model is
}
