% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_sig_test_nlme.R
\name{.robust_sig_test_nlme}
\alias{.robust_sig_test_nlme}
\title{Perform heteroskedasticity robust significance tests}
\usage{
.robust_sig_test_nlme(object, type = "HC1")
}
\arguments{
\item{object}{Model object of type "nlme"}
}
\description{
Performs heteroskedasticity-robust t-tests for coefficients in a (generalized) linear model.
}
