% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_vcov_lm_lme4.R
\name{.robust_vcov_lm_lme4}
\alias{.robust_vcov_lm_lme4}
\title{Compute HC variance-covariance matrix}
\usage{
.robust_vcov_lm_lme4(object)
}
\arguments{
\item{object}{Model object of type "lm", "glm" or "lmer"(lme4)}
}
\value{
Covariance matrix
}
\description{
Computes heteroskedasticity robust covariance matrix
}
