% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_vcov_nlme.R
\name{.robust_vcov_nlme}
\alias{.robust_vcov_nlme}
\title{Compute HC variance-covariance matrix}
\usage{
.robust_vcov_nlme(object)
}
\arguments{
\item{object}{Model object of type "nlme" or "lme"}
}
\value{
Covariance matrix
}
\description{
Computes heteroskedasticity robust covariance matrix
}
