% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnson_neyman_wrapper.R
\name{johnson_neyman}
\alias{johnson_neyman}
\title{Generate Johnson-Neyman plot}
\usage{
johnson_neyman(object, X, M, modrange = NULL, resolution = 10000)
}
\arguments{
\item{object}{model object, either "lm","lmer","nlme"}

\item{X}{name of independent variable}

\item{M}{name of moderator variable}

\item{modrange}{moderator range to consider. Defaults to observed range.}

\item{resolution}{plotting resolution, i.e. count of moderator values to consider}
}
\value{
Johnson-neyman plot as gpplot2 object.
}
\description{
Generates a Johnson-Neyman plot for the effect of X on the dependent variable, moderated by M.
}
\examples{
x <- rnorm(50)
m <- rnorm(50)
y <- rnorm(50,x+m,2)
model <- lm(y~x*m)
plot <- johnson_neyman(model,"x","m")
plot + labs(x="Changed x label",y="Changed y label")

}
