% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_sig_test.R
\name{robust_sig_test}
\alias{robust_sig_test}
\title{Perform heteroskedasticity robust significance tests}
\usage{
robust_sig_test(object, type = "HC1")
}
\arguments{
\item{object}{Model object of type "lm", "glm", "lmer", "glmer"}
}
\description{
Performs heteroskedasticity-robust t-tests for coefficients in a (generalized) linear (mixed) model.
}
