% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_sig_test_lm.R
\name{robust_sig_test_lm}
\alias{robust_sig_test_lm}
\title{Perform heteroskedasticity robust significance tests}
\usage{
robust_sig_test_lm(object)
}
\arguments{
\item{object}{Model object of type "lm" or "glm"}
}
\description{
Performs heteroskedasticity-robust t-tests for coefficients in a (generalized) linear model.
}
