% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_sig_test_lme4.R
\name{robust_sig_test_lme4}
\alias{robust_sig_test_lme4}
\title{Perform heteroskedasticity robust significance tests}
\usage{
robust_sig_test_lme4(object)
}
\arguments{
\item{object}{Model object of type "lmer"(lme4)}
}
\description{
Performs heteroskedasticity-robust t-tests for coefficients in a linear mixed model.
}
