% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_vcov.R
\name{robust_vcov}
\alias{robust_vcov}
\title{Compute HC variance-covariance matrix}
\usage{
robust_vcov(object)
}
\arguments{
\item{object}{Model object of type "lm", "glm" or "lmer"(lme4)}
}
\value{
Covariance matrix
}
\description{
Computes heteroskedasticity robust covariance matrix
}
