% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootreg.R
\name{bootreg}
\alias{bootreg}
\title{Run Bootstrapped regression}
\usage{
bootreg(data, model, method = lm, n = 1000)
}
\arguments{
\item{data}{a data frame}

\item{model}{a fitted lme or lm - model}

\item{method}{a function to use for refitting}

\item{n}{number of bootstrap samples}
}
\value{
Bootstrap regression results object (brr)
}
\description{
runs repeated regressions on bootstrap samples, testing coefficients via z-distribution using bootstrapped standard errors.
}
\examples{
R base models
x <- rnorm(100)
m <- rnorm(100)
y <- rnorm(100,0.4*x-0.6*m,2)
model <- lm(y~m*x)
brr <- bootreg(data.frame(x,m,y),model,lm)


#mixed models
x <- rnorm(50)
m <- rnorm(50)
y <- rnorm(50,x*m,2)
g <- sample(1:10,50,TRUE)
model <- lme4::lmer(y~m*x + (1|g))
brr <- bootreg(data.frame(x,m,y,g),model,lme4::lmer)

}
