% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_vcov.R
\name{robust_vcov}
\alias{robust_vcov}
\title{Compute HC variance-covariance matrix}
\usage{
robust_vcov(object, modelmat = NULL, type = "HC1")
}
\arguments{
\item{object}{Model object of type "lm", "glm", "lmer", "glmer", "nlme", "lme"}

\item{modelmat}{Model matrix, required only if model is of none of the allowed classes}

\item{type}{Any of "HC1", "HC3", indicating method for calculating HC variance-covariance matrix}
}
\value{
Covariance matrix
}
\description{
Computes heteroskedasticity robust covariance matrix
}
